+++ packages/rpc/src/paramsFormatter.ts
@@ -164,6 +164,17 @@ export const formatter = {
       ...rest,
     };
   },
+
+  toTransactionAndWitnessProof: (
+    proof: CKBComponents.TransactionAndWitnessProof
+  ): RPC.TransactionAndWitnessProof => {
+    return {
+      block_hash: proof.blockHash,
+      witnesses_proof: proof.witnessesProof,
+      transactions_proof: proof.transactionsProof,
+    };
+  },
+
   toSearchFilter: (data: CKBComponents.SearchFilter): RPC.SearchFilter => {
     if (!data) return data;
     return {
@@ -180,6 +191,9 @@ export const formatter = {
       script: formatter.toScript(data.script),
       script_type: data.scriptType,
       filter: data.filter ? formatter.toSearchFilter(data.filter) : data.filter,
+      script_search_mode: data.scriptSearchMode
+        ? data.scriptSearchMode
+        : "prefix",
     };
   },
   toGetCellsSearchKey: (