+++ packages/rpc/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.20.0-alpha.2](https://github.com/ckb-js/lumos/compare/v0.19.0...v0.20.0-alpha.2) (2023-04-04)
+
+
+### Features
+
+* **ckb-indexer:** fix loadCells DoS vuln. Support the exact search mode ([#491](https://github.com/ckb-js/lumos/issues/491)) ([e73ca05](https://github.com/ckb-js/lumos/commit/e73ca0502b0feb3e98be848f76325a912052d0f0))
+* **rpc:** support rpc runs on service worker ([#493](https://github.com/ckb-js/lumos/issues/493)) ([72830ef](https://github.com/ckb-js/lumos/commit/72830ef3eb85d49b3b77bc37ecbdc58ada2d380c))
+
+
+
+
+
 # [0.20.0-alpha.1](https://github.com/ckb-js/lumos/compare/v0.19.0...v0.20.0-alpha.1) (2022-12-01)
 
 **Note:** Version bump only for package @ckb-lumos/rpc